explorer,views: fix charts page target ticket pool size
In (*explorerUI).Charts get target pool size from
exp.pageData.HomeInfo.PoolInfo.Target, and pass to page template.

explorer/types: change TicketPoolInfo.Target uint16=>uint32

+++ explorer/types/explorertypes.go
@@ -859,7 +859,7 @@ type TicketPoolInfo struct {
 	Value         float64 `json:"value"`
 	ValAvg        float64 `json:"valavg"`
 	Percentage    float64 `json:"percent"`
-	Target        uint16  `json:"target"`
+	Target        uint32  `json:"target"`
 	PercentTarget float64 `json:"percent_target"`
 }
